requirements.txt
numpy>=1.21.0
torch>=1.9.0
pygame>=2.0.1
matplotlib>=3.4.0
tqdm>=4.62.0
PyYAML>=5.4.1
scipy>=1.8.0
opencv-python>=4.5.0
shapely>=1.8.0
casadi>=3.5.5 